+++ DotGame/src/globals/globals.h
@@ -27,12 +27,12 @@ extern float gMouseY;
 const int SCR_WIDTH  = 1280;
 const int SCR_HEIGHT = 720;
 
-const int FRAME_WIDTH  = 64;
-const int FRAME_HEIGHT = 64;
-const int NUM_COLS     = 10;
-const int NUM_ROWS     = 10;
-const int GRID_POS_X   = 100;
-const int GRID_POS_Y   = 50;
+const float FRAME_WIDTH  = 64;
+const float FRAME_HEIGHT = 64;
+const int   NUM_COLS     = 10;
+const int   NUM_ROWS     = 10;
+const int   GRID_POS_X   = 100;
+const int   GRID_POS_Y   = 50;
 
 const std::string DATA_FOLDER = "data/";
